fix bugs
deleted duplicated objects that were not good
