+++ app/src/main/AndroidManifest.xml
@@ -11,18 +11,18 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar"
         tools:ignore="GoogleAppIndexingWarning">
-        <activity android:name=".OrderConfirmationActivity"></activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".RegisterActivity" />
-        <activity android:name=".ManaActivity" />
-        <activity android:name=".ManaPickerActivity" />
-        <activity android:name=".MainActivity" />
+        <activity android:name=".OrderConfirmationActivity" />
+        <activity android:name=".LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".RegisterActivity" />
+        <activity android:name=".ManaActivity" />
+        <activity android:name=".ManaPickerActivity" />
+        <activity android:name=".MainActivity"/>
+
 
         <meta-data
             android:name="preloaded_fonts"
